Moved providers to own folder & cleaned up App.tsx a bit
