+++ frontend/client/src/providers/project/project.provider.tsx
@@ -1,4 +1,7 @@
+/* Lib imports */
 import { createContext, ReactNode, useContext, useMemo, useState } from "react";
+
+/* Types */
 import { Project } from "../../types/Project";
 
 type ProjectProviderType = {
